Untangle post parameters parsing:
The post parameters parsing was awkward. The "_body_params"
attribute would only return values that were created by the
_set_body_params() method. That method was run by the
_parse_post_params() method (which would first check the
"_body_params" attribute. That's basically what a lazy attribute
is for. The _parse_post_params() is called by the _build_uploads(),
if the "is_body_parsed" is not set.

Now we do it differently:

_build_uploads() method calls "_body_params" attribute. That
attribute makes sure the body has been parsed by calling the "body"
attribute. Then it builds the body parameters the same way (and
calls the "body_is_parsed" attribute while at it).