Introduce Hash::MultiValue parameters + keywords:
Finally we get to the gist of it. We're introducing several
keywords and a new type of value.

The following keywords are added:

* query_parameters: Query parameters, akin to params('query')
* body_parameters: Body parameters, akin to params('body')
* route_parameters: Route parameters, akin to params('route')

They do not return the values, but instead an object of
Hash::MultiValue. You can then use this object to fetch the
parameter values.

    get '/:foo' => sub {
        my $foo = route_parameters->get('foo');
        ...
    };

    post '/' => sub {
        my @foo = body_parameters->get_all('foo');
        ...
    };

They also work for serialized data, if they are of a hash.

We explicitly do not have a DSL for all merged parameters, since
we consider that a bad practice (and will start warning against it
at some point probably), but you *can* reach them using the request
object:

    request->parameters->get('bar'); # merged parameters' "bar"