Handle capture parameters:
The capture parameters receive the same treeatment as the splat
parameters, which means they get removed as well, and it will
eventually become a saved key. "route_parameters" will not show
it.

Additionally, the tests have exposed that while we do return an
array for splat() all the time, the captures() keyword does not
always return a hashref. if there were no captures, it returns
undef. That's bad. Once we fix it, the test will show it and we
will update it.